+++ internal/engine/synclet_build_and_deployer.go
@@ -10,7 +10,9 @@ import (
 
 var _ BuildAndDeployer = &SyncletBuildAndDeployer{}
 
-type SyncletBuildAndDeployer struct{}
+type SyncletBuildAndDeployer struct {
+	scli *Client
+}
 
 func NewSyncletBuildAndDeployer() *SyncletBuildAndDeployer {
 	return &SyncletBuildAndDeployer{}
